requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
anthropic==0.39.0
python-dotenv==1.0.0
schedule==1.2.0
